# Use memcpy() of standard C library