@Watson1978 2022-01-15
Here is stacktrace with before changing.

<img width="956" alt="スクリーンショット 2022-01-16 0 18 06" src="https://user-images.githubusercontent.com/199156/149627029-c463b70f-4366-4982-9ca8-a9d42513d85a.png">